// src/server/models/User.ts
export interface IUser extends Document {
    /** Full user name. */
    fullName: string;

    /** User email ID. */
    email: string;

    /** Password (empty for OAuth users). */
    password?: string;

    /** User creation date. */
    createdAt?: Date;

    /** Refresh token for the user. */
    refreshToken?: string;

    /** Associated OAuth accounts */
    oAuthProfiles: [{
        /** The provider with which the user authenticated */
        provider: string,
        /** A unique identifier for the user, as generated by the provider */
        id: string,
    }];

    /**
     * Check password of this user.
     *
     * @async
     *
     * @param password The password to check.
     *
     * @returns
     *  A promise which resolves to <code>true</code> if the passwords match or
     *  <code>false<code> if the passwords do not match.
     */
    checkPassword(password: string): Promise<boolean>;

    /**
     * Generate a new access token for the user.
     *
     * @returns A promise which resolves to the generated access token.
     */
    generateAccessToken(): Promise<string>;

    /**
     * Generate and save a new refresh token for the user.
     *
     * @returns A promise which resolves to the generated refresh token.
     */
    generateRefreshToken(): Promise<string>;
}

const ACCESS_TOKEN_EXPIRE = 60 * 5;
const ACCESS_TOKEN_SECRET = process.env.ACCESS_TOKEN_SECRET || 'secret';
const REFRESH_TOKEN_EXPIRE = 60 * 60 * 24 * 7;
const REFRESH_TOKEN_SECRET = process.env.REFRESH_TOKEN_SECRET || 'secret';
const SALT_ROUNDS = 10;

const userSchema = new Schema({
    fullName: {
        type: String,
        required: true,
    },
    email: {
        type: String,
        required: true,
        unique: true,
        min: 3,
        max: 320,
    },
    password: {
        type: String,
        min: 6,
        max: 1024,
    },
    createdAt: {
        type: Date,
        default: Date.now,
    },
    refreshToken: {
        type: String,
    },
    oAuthProfiles: [{
        provider: {
            type: String,
            required: true,
        },
        id: {
            type: String,
            required: true,
        },
    }],
});
// ...
export const User = model<IUser>('User', userSchema);